Updated test script to pass string to mocha
no-issue

This is because the shell will try to expand the glob if it's not
wrapped in strings, which can give different outcomes depending on the
globbing supporting by the shell. Passing a string will allow mocha to
sort out globbing which is more predictable.

Related issue is here: https://github.com/mochajs/mocha/issues/1828